Refactor home page structure: remove `EnsTest`, inline `HomePage` logic into `index.js`, update imports, and add `/ens-test` redirect.
