+++ pages/index.js
@@ -1,19 +1,20 @@
-import { Layout, LayoutContent } from "@uoguelph/react-components/layout";
+import { getSpotlightCards, getSpotlightHero } from "@/data/drupal/home";
+import { twJoin } from "tailwind-merge";
+import { Meta } from "@/components/meta";
+import AppArmor from "@/components/app-armor";
 import { Header } from "@uoguelph/react-components/header";
-import { Footer } from "@uoguelph/react-components/footer";
+import { Layout, LayoutContent } from "@uoguelph/react-components/layout";
+import { TagLine } from "@/components/home/tag-line";
+import { SpotlightHero } from "@/components/home/spotlight-hero";
 import { Container } from "@uoguelph/react-components/container";
 import { Typography } from "@uoguelph/react-components/typography";
 import { SpotlightCards } from "@/components/home/spotlight-cards";
-import { TagLine } from "@/components/home/tag-line";
+import { StudyHere } from "@/components/home/study-here";
 import { Rankings } from "@/components/home/rankings";
 import { ThreeCampuses } from "@/components/home/three-campuses";
-import { SpotlightHero } from "@/components/home/spotlight-hero";
 import { HomeStory } from "@/components/home/story";
-import { StudyHere } from "@/components/home/study-here";
-import { getSpotlightCards, getSpotlightHero } from "@/data/drupal/home";
-import { twJoin } from "tailwind-merge";
-import AppArmor from "@/components/app-armor";
-import { HomePage } from "@/components/home/page";
+import { Footer } from "@uoguelph/react-components/footer";
+import { useRouter } from "next/router";
 
 export async function getStaticProps(context) {
   const status = context?.preview || process.env.NODE_ENV !== "production" ? null : true;
@@ -29,5 +30,59 @@ export async function getStaticProps(context) {
 }
 
 export default function Home({ cards, hero }) {
-  return <HomePage cards={cards} hero={hero} />;
+  const containerClasses = twJoin("pt-6");
+  const { query } = useRouter();
+
+  console.log(query);
+
+  return (
+    <>
+      <Meta></Meta>
+
+      <AppArmor test={query["ens-test"] === "true"} />
+
+      <Header></Header>
+
+      <Layout>
+        <LayoutContent container={false}>
+          <div className="flex flex-col-reverse">
+            <TagLine />
+            {hero && <SpotlightHero data={hero} />}
+          </div>
+
+          <Container>
+            <Typography className="text-black!" type="h2" as="h2">
+              Our Latest News and Events
+            </Typography>
+            <SpotlightCards cards={cards} />
+          </Container>
+
+          <Container className={containerClasses}>
+            <Typography className="text-black!" type="h2" as="h2">
+              Study Here
+            </Typography>
+            <StudyHere />
+          </Container>
+
+          <Container className={containerClasses}>
+            <Typography className="text-black!" type="h2" as="h2">
+              How We Rank Among the World
+            </Typography>
+            <Rankings />
+          </Container>
+
+          <Container className={containerClasses}>
+            <Typography className="text-black!" type="h2" as="h2">
+              Our Three Campuses
+            </Typography>
+            <ThreeCampuses />
+          </Container>
+
+          <HomeStory />
+        </LayoutContent>
+      </Layout>
+
+      <Footer></Footer>
+    </>
+  );
 }
